requirements.txt
Flask
Flask-Session
Flask-JSGlue
requests
pymongo
dropbox

# local metadata extraction
epub_meta
pdfrw

# file-type recognition
python-magic

# image handling
Pillow
